+++ src/solvers/exact_penalty_method.jl
@@ -301,7 +301,7 @@ function exact_penalty_method!(
     smoothing=LogarithmicSumOfExponentials(),
     sub_cost=ExactPenaltyCost(cmo, ρ, u; smoothing=smoothing),
     sub_grad=ExactPenaltyGrad(cmo, ρ, u; smoothing=smoothing),
-    sub_kwargs=[],
+    sub_kwargs=(;),
     sub_problem::AbstractManoptProblem=DefaultManoptProblem(
         M,
         decorate_objective!(
@@ -324,6 +324,7 @@ function exact_penalty_method!(
             ),
             stopping_criterion=sub_stopping_criterion,
             stepsize=default_stepsize(M, QuasiNewtonState),
+            sub_kwargs...,
         );
         sub_kwargs...,
     ),